requirements.txt
absl-py==0.9.0
astor==0.8.1
audioread==2.1.8
certifi==2020.4.5.2
cffi==1.14.0
click==7.1.2
decorator==4.4.2
ffmpeg==1.4
Flask==1.1.1
gast==0.3.3
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
hdf5storage==0.1.15
importlib-metadata==1.6.1
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
librosa==0.7.1
llvmlite==0.31.0
Markdown==3.2.2
MarkupSafe==1.1.1
numba==0.46.0
numpy==1.18.5
Pillow==7.1.2
protobuf==3.12.2
pycparser==2.20
pydub==0.24.1
resampy==0.2.2
scikit-learn==0.23.1
scipy==1.3.2
six==1.15.0
SoundFile==0.10.3.post1
SpeechRecognition==3.8.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
TextGrid==1.5
threadpoolctl==2.1.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.46.1
Unidecode==1.1.1
webrtcvad==2.0.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0